requirements.txt
pandas==2.1.1
python-dotenv==1.0.0
langchain
langchain-groq
fastapi
uvicorn
matplotlib
python-multipart
gunicorn